// src/services/generated/models/aCLType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export type ACLType = typeof ACLType[keyof typeof ACLType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ACLType = {
  GLOBAL: 'GLOBAL',
  PRIVATE: 'PRIVATE',
} as const;